clean up test setup of bundle .tests.ecmatestsuite
