+++ tests/org.eclipse.n4js.tests.ecmatestsuite/src/org/eclipse/n4js/n4jsx/ecmatestsuite/ECMA5TestInSuiteForJSX.java
@@ -15,7 +15,7 @@ import org.eclipse.n4js.JSLibSingleTestConfig;
 import org.eclipse.n4js.N4JSInjectorProvider;
 import org.eclipse.n4js.analysis.Analyser;
 import org.eclipse.n4js.n4JS.Script;
-import org.eclipse.n4js.tests.ecmatestsuite.ECMA5TestSuite;
+import org.eclipse.n4js.tests.ecmatestsuite.ECMA5TestInSuite;
 import org.eclipse.xtext.resource.XtextResourceSet;
 import org.eclipse.xtext.testing.InjectWith;
 
@@ -23,7 +23,7 @@ import org.eclipse.xtext.testing.InjectWith;
  *
  */
 @InjectWith(N4JSInjectorProvider.class)
-public class ECMA5TestSuiteForJSX extends ECMA5TestSuite {
+public class ECMA5TestInSuiteForJSX extends ECMA5TestInSuite {
 
 	@Override
 	protected Script doParse(String code, URI uri, XtextResourceSet resourceSet, Analyser analyser) throws Exception {
@@ -43,7 +43,7 @@ public class ECMA5TestSuiteForJSX extends ECMA5TestSuite {
 	/**
 	 * Constructor
 	 */
-	public ECMA5TestSuiteForJSX(JSLibSingleTestConfig config) {
+	public ECMA5TestInSuiteForJSX(JSLibSingleTestConfig config) {
 		super(config);
 	}
